*Added 'delete load' button and functionality for this
*Moved loads post and delete handlers to separate files
