+++ homework3/semi-uber/UI/src/app/shipper/shipper.component.html
@@ -36,6 +36,9 @@
 <ul class="loads">
   <li *ngFor="let load of userLoads" class="loads__item load-card">
     <div class="load-card__title">Load id: {{ load._id }}</div>
+    <div class="load-card__asigned">
+      <span class="load-card__prop">Assigned to: </span>{{ load.assigned_to }}
+    </div>
     <div class="load-card__status">
       <span class="load-card__prop">Status: </span>{{ load.status }}
     </div>
@@ -61,7 +64,9 @@
     <div class="load-card__buttons">
       <button class="load-card__update">update</button>
       <button class="load-card__post">post</button>
-      <button class="load-card__delete" (click)="deleteLoad(load)">delete</button>
+      <button class="load-card__delete" (click)="deleteLoad(load)">
+        delete
+      </button>
     </div>
   </li>
 </ul>
