+++ homework3/semi-uber/UI/src/app/shipper/shipper.component.scss
@@ -40,14 +40,20 @@
   &__prop {
     font-weight: 700;
   }
-  &__status, &__state, &__dimensions, &__payload {
+  &__status,
+  &__state,
+  &__dimensions,
+  &__payload,
+  &__asigned {
     margin: 10px 0;
   }
   &__buttons {
-    width: 100%
+    width: 100%;
   }
-  &__update,&__post, &__delete {
+  &__update,
+  &__post,
+  &__delete {
     text-transform: uppercase;
-    margin:  5px;
+    margin: 5px;
   }
 }
